Redesign RDE project page with HADES and DEIMOS profiles
